[REF] rename resnets_examples.py to resnet_cases.py
