+++ test/extensions/secondorder/diag_ggn/diag_ggn_settings.py
@@ -11,7 +11,7 @@ Additional local cases can be defined here through ``LOCAL_SETTINGS``.
 """
 from test.core.derivatives.utils import classification_targets, regression_targets
 from test.extensions.secondorder.secondorder_settings import SECONDORDER_SETTINGS
-from test.resnet.resnets_examples import ResNet1, ResNet2
+from test.resnet.resnet_cases import ResNet1, ResNet2
 from test.utils.evaluation_mode import initialize_training_false_recursive
 
 from torch import rand
